backend/requirements.txt
# API Framework
fastapi==0.102.0
uvicorn[standard]==0.23.2
python-multipart>=0.0.6

# Environment
python-dotenv==1.0.1

# Google Cloud Vision OCR
google-cloud-vision==3.12.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Authentication
pyjwt[crypto]>=2.8.0
bcrypt>=4.1.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Image Preprocessing
opencv-python>=4.8.0
numpy>=1.24.0

#AI
google-genai